separate recovery code generation from recoveryauth
