+++ backend/authschemes/recoveryauth/services.go
@@ -5,14 +5,13 @@ package recoveryauth
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/hex"
 	"errors"
 	"time"
 
 	"github.com/theparanoids/ashirt-server/backend"
 	"github.com/theparanoids/ashirt-server/backend/authschemes"
 	recoveryConsts "github.com/theparanoids/ashirt-server/backend/authschemes/recoveryauth/constants"
+	recoveryHelpers "github.com/theparanoids/ashirt-server/backend/authschemes/recoveryauth/helpers"
 	"github.com/theparanoids/ashirt-server/backend/database"
 	"github.com/theparanoids/ashirt-server/backend/policy"
 	"github.com/theparanoids/ashirt-server/backend/server/middleware"
@@ -49,7 +48,7 @@ func generateRecoveryCodeForUser(ctx context.Context, bridge authschemes.AShirtA
 		return nil, backend.WrapError("Unable to get UserID from slug", err)
 	}
 
-	authKeyStr, err := GenerateRecoveryCodeForUser(bridge.GetDatabase(), userID)
+	authKeyStr, err := recoveryHelpers.GenerateRecoveryCodeForUser(bridge.GetDatabase(), userID)
 	if err != nil {
 		return nil, backend.WrapError("Could not generate recovery code for user", err)
 	}
@@ -106,29 +105,4 @@ func generateRecoveryEmail(ctx context.Context, bridge authschemes.AShirtAuthBri
 	}
 
 	return nil
-}
-
-// GenerateRecoveryCodeForUser provides a mechanism for non-recovery services to generate and register
-// a recovery authentication for the provided UserID.
-//
-// returns the generatedCode if successful, and an error, if one was encountered
-func GenerateRecoveryCodeForUser(db *database.Connection, userID int64) (string, error) {
-	authKey := make([]byte, recoveryKeyLength)
-	if _, err := rand.Read(authKey); err != nil {
-		return "", backend.WrapError("Unable to generate random recovery key", err)
-	}
-	authKeyStr := hex.EncodeToString(authKey)
-
-	err := GenerateNewRecoveryRecord(db, authschemes.UserAuthData{
-		UserID:  userID,
-		UserKey: authKeyStr,
-	})
-
-	return authKeyStr, err
-}
-
-// GenerateNewRecoveryRecord is a shorthand for CreateNewAuthForUserGeneric with the recovery code
-// constant provide
-func GenerateNewRecoveryRecord(db *database.Connection, userAuthData authschemes.UserAuthData) error {
-	return authschemes.CreateNewAuthForUserGeneric(db, recoveryConsts.Code, userAuthData)
 }
